Make the assets/themes folder if it doesn't exist
- Also only run clearThemes if the folder exists

+++ styles/src/buildThemes.ts
@@ -17,11 +17,15 @@ const tempDirectory = fs.mkdtempSync(path.join(tmpdir(), "build-themes"));
 
 // Clear existing themes
 function clearThemes(themeDirectory: string) {
-  for (const file of fs.readdirSync(themeDirectory)) {
-    if (file.endsWith(".json")) {
-      const name = file.replace(/\.json$/, "");
-      if (!colorSchemes.find((colorScheme) => colorScheme.name === name)) {
-        fs.unlinkSync(path.join(themeDirectory, file));
+  if (!fs.existsSync(themeDirectory)) {
+    fs.mkdirSync(themeDirectory, { recursive: true });
+  } else {
+    for (const file of fs.readdirSync(themeDirectory)) {
+      if (file.endsWith(".json")) {
+        const name = file.replace(/\.json$/, "");
+        if (!colorSchemes.find((colorScheme) => colorScheme.name === name)) {
+          fs.unlinkSync(path.join(themeDirectory, file));
+        }
       }
     }
   }
